+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     * Various fixes to the SVG parser
     * Added support for Android Vector Drawables
     * Improved support for animated SVG paths
+    * Opening raster images uses the file basename as layer name
 
 ## 0.5.2